+++ lib/entry/v2/EntryUpdateCL.ts
@@ -31,10 +31,14 @@ export default class EntryUpdateCL<EntityT extends object = object, EntityKeysT 
         this.getODataModel().setDefaultBindingMode("TwoWay");
         this.getODataModel().setUseBatch(true);
 
-        if (this.getDialogStrategy() === DialogStrategies.LOAD) {
-            await this.loadDialog();
+        if (this.getDisplayObjectPage()) {
+            await this.createObjectPage();
         } else {
-            await this.createDialog();
+            if (this.getDialogStrategy() === DialogStrategies.LOAD) {
+                await this.loadDialog();
+            } else {
+                await this.createDialog();
+            }
         }
     }
 
@@ -82,7 +86,7 @@ export default class EntryUpdateCL<EntityT extends object = object, EntityKeysT 
         }
 
         this.submit();
-    }    
+    }
 
     private onEntryCanceled(event: Button$PressEvent) {
         this.reset();
@@ -118,6 +122,38 @@ export default class EntryUpdateCL<EntityT extends object = object, EntityKeysT 
         } else {
             fragment.destroyFragmentContent();
             throw new Error("Provided fragment must contain a sap.m.Dialog control. Put all the controls into a sap.m.Dialog");
-        }        
+        }
+    }
+
+    private async createObjectPage() {
+        await this.initializeContext(this.settings.initializer);
+        const content = new ContentCL<EntryUpdateCL<EntityT>, EntityT>(this.getSourceController(), this, ODataMethods.UPDATE, this.getModelName());
+
+        // Create Object Page
+        this.createObjectPageLayout();
+        const objectPageInstance = this.getObjectPageInstance();
+        objectPageInstance.addCompleteButton(this.getBeginButtonText(), this.getBeginButtonType());
+        objectPageInstance.addCancelButton(this.getEndButtonText(), this.getEndButtonType());
+
+        if (this.getFormType() === FormTypes.SMART) {
+            await content.addSmartSections();
+            objectPageInstance.getObjectPageLayout().setModel(this.getODataModel());
+            objectPageInstance.getObjectPageLayout().setBindingContext(this.getEntryContext());
+        } else {
+            await content.addSimpleSections();
+            objectPageInstance.getObjectPageLayout().setModel(this.getODataModel(), this.getModelName());
+            objectPageInstance.getObjectPageLayout().setBindingContext(this.getEntryContext(), this.getModelName());
+        }
+
+        if (this.getCustomContents().length) {
+            objectPageInstance.addEmptySection(this.getCustomContentSectionTitle());
+
+            this.getCustomContents().forEach((customContent) => {
+                objectPageInstance.addContentToSection(customContent);
+            });
+        }
+
+        await this.createTypedView();
+        this.displayTypedView();
     }
 }